+++ changelog.md
@@ -9,6 +9,7 @@
 * [changed] Generic Locality List view used
 * [added] Ordering in identities
 * [added] Merge multiple identities
+* [added] Suggest code extraction from the name of identity 
 
 ## 01-2025
 * [added] Added carousel for media files in the detail view